% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{build_graph}
\alias{build_graph}
\title{Build a directed igraph from nodes/edges.}
\usage{
build_graph(nodes, edges)
}
\arguments{
\item{nodes}{tibble with \code{openalex_id} as first column (vertex names).}

\item{edges}{tibble with \code{source_id}, \code{target_id}.}
}
\value{
igraph directed graph.
}
\description{
Build a directed igraph from nodes/edges.
}
