% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chunked}
\alias{chunked}
\title{Split a vector into chunks of at most \code{size} elements.}
\usage{
chunked(items, size)
}
\arguments{
\item{items}{vector}

\item{size}{integer > 0}
}
\value{
list of vectors, each of length \code{size} (last may be shorter).
}
\description{
Split a vector into chunks of at most \code{size} elements.
}
