% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_pmid}
\alias{clean_pmid}
\title{Extract numeric PMID.
Rejects \code{"0"}, \code{"0.0"}, \code{"unavailable"}, \code{"n/a"}, \code{"na"}.}
\usage{
clean_pmid(value)
}
\arguments{
\item{value}{any R object}
}
\value{
PMID string or \code{NA_character_}.
}
\description{
Extract numeric PMID.
Rejects \code{"0"}, \code{"0.0"}, \code{"unavailable"}, \code{"n/a"}, \code{"na"}.
}
