% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cmd_report}
\alias{cmd_report}
\title{Write the report markdown.}
\usage{
cmd_report(args, settings)
}
\arguments{
\item{args}{argparse namespace (unused).}

\item{settings}{\code{Settings} instance.}
}
\value{
integer exit code.
}
\description{
Write the report markdown.
}
