% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compact_openalex_id}
\alias{compact_openalex_id}
\title{Compact an OpenAlex ID by stripping the \verb{https://openalex.org/} prefix.}
\usage{
compact_openalex_id(value)
}
\arguments{
\item{value}{any R object}
}
\value{
short ID (e.g. \code{"W123456"}) or \code{NA_character_}.
}
\description{
Compact an OpenAlex ID by stripping the \verb{https://openalex.org/} prefix.
}
