% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{df_to_records}
\alias{df_to_records}
\title{Convert a data frame into a list of named lists (one per row).}
\usage{
df_to_records(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list of named lists.
}
\description{
Convert a data frame into a list of named lists (one per row).
}
