% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations.R
\name{doi_from_node_id}
\alias{doi_from_node_id}
\title{Extract the bare DOI from a \verb{doi:<doi>} node ID.}
\usage{
doi_from_node_id(node_id)
}
\arguments{
\item{node_id}{character (possibly \code{NA}).}
}
\value{
bare lowercase DOI or \code{NA_character_}.
}
\description{
Extract the bare DOI from a \verb{doi:<doi>} node ID.
}
