% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{doi_url}
\alias{doi_url}
\title{Build a DOI URL from a value.}
\usage{
doi_url(value)
}
\arguments{
\item{value}{any R object}
}
\value{
\verb{https://doi.org/<doi>} or \code{NA_character_}.
}
\description{
Build a DOI URL from a value.
}
