% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Create a directory (recursively) if needed.}
\usage{
ensure_dir(path)
}
\arguments{
\item{path}{character}
}
\value{
the path, invisibly.
}
\description{
Create a directory (recursively) if needed.
}
