% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{first_author_last_name}
\alias{first_author_last_name}
\title{Extract the last name of the first author from a semicolon-separated field.}
\usage{
first_author_last_name(author_field)
}
\arguments{
\item{author_field}{character}
}
\value{
last name string or \code{NA_character_}.
}
\description{
Extract the last name of the first author from a semicolon-separated field.
}
