% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{full_openalex_id}
\alias{full_openalex_id}
\title{Expand a compact OpenAlex ID to the full \verb{https://openalex.org/<id>} form.}
\usage{
full_openalex_id(value)
}
\arguments{
\item{value}{any R object}
}
\value{
full URL or \code{NA_character_}.
}
\description{
Expand a compact OpenAlex ID to the full \verb{https://openalex.org/<id>} form.
}
