% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_missing}
\alias{is_missing}
\title{Missingness check}
\usage{
is_missing(value)
}
\arguments{
\item{value}{any R object}
}
\value{
\code{TRUE} if value is \code{NA}, \code{NaN}, or length 0, else \code{FALSE}.
}
\description{
Missingness check
}
