% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations_pipeline.R
\name{is_seed_id}
\alias{is_seed_id}
\title{Quick SQL check: is the given node_id also a seed openalex_id?}
\usage{
is_seed_id(store, node_id)
}
\arguments{
\item{store}{StudyStore instance.}

\item{node_id}{character node ID.}
}
\value{
logical.
}
\description{
Quick SQL check: is the given node_id also a seed openalex_id?
}
