% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{is_seed_notice}
\alias{is_seed_notice}
\title{Is the RetractionNature value a seed notice (retraction or EOC)?}
\usage{
is_seed_notice(nature)
}
\arguments{
\item{nature}{character (possibly \code{NA}).}
}
\value{
logical.
}
\description{
Is the RetractionNature value a seed notice (retraction or EOC)?
}
