% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{json_dumps}
\alias{json_dumps}
\title{Serialize an R object to a JSON string.}
\usage{
json_dumps(value)
}
\arguments{
\item{value}{any R object}
}
\value{
JSON character string.
}
\description{
Serialize an R object to a JSON string.
}
