% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{make_openalex}
\alias{make_openalex}
\title{Create an \code{OpenAlexClient} from settings.}
\usage{
make_openalex(settings)
}
\arguments{
\item{settings}{\code{Settings} instance.}
}
\value{
An \code{OpenAlexClient} instance.
}
\description{
Create an \code{OpenAlexClient} from settings.
}
