% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{make_opencitations}
\alias{make_opencitations}
\title{Create an \code{OpenCitationsClient} from settings.}
\usage{
make_opencitations(settings)
}
\arguments{
\item{settings}{\code{Settings} instance.}
}
\value{
An \code{OpenCitationsClient} instance.
}
\description{
Create an \code{OpenCitationsClient} from settings.
}
