% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.R
\name{manuscript_save}
\alias{manuscript_save}
\title{Save a ggplot as both PNG and SVG.}
\usage{
manuscript_save(plot, path_no_ext, width = 8, height = 5)
}
\arguments{
\item{plot}{ggplot.}

\item{path_no_ext}{character path without extension.}

\item{width, height}{numeric inches.}
}
\value{
invisible NULL.
}
\description{
Save a ggplot as both PNG and SVG.
}
