% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_units.R
\name{min_date_or_na}
\alias{min_date_or_na}
\title{Return the minimum Date or \code{NA}.}
\usage{
min_date_or_na(values)
}
\arguments{
\item{values}{vector coercible to Date.}
}
\value{
Date.
}
\description{
Return the minimum Date or \code{NA}.
}
