% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex.R
\name{new_openalex_error}
\alias{new_openalex_error}
\title{Construct an \code{OpenAlexError} condition object (without raising).}
\usage{
new_openalex_error(
  msg,
  status_code = NA_integer_,
  response_text = NA_character_
)
}
\arguments{
\item{msg}{character message.}

\item{status_code}{integerish HTTP status, or \code{NA}.}

\item{response_text}{character response body, or \code{NA}.}
}
\value{
An object of class \code{c("OpenAlexError", "error", "condition")}.
}
\description{
Construct an \code{OpenAlexError} condition object (without raising).
}
