% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations.R
\name{open_citations_error}
\alias{open_citations_error}
\title{Construct an \code{OpenCitationsError} condition object.}
\usage{
open_citations_error(msg)
}
\arguments{
\item{msg}{character message.}
}
\value{
An object of class \code{c("OpenCitationsError", "error", "condition")}.
}
\description{
Construct an \code{OpenCitationsError} condition object.
}
