% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openalex.R
\name{openalex_url}
\alias{openalex_url}
\title{Build the full OpenAlex URL for a compact ID.}
\usage{
openalex_url(openalex_id)
}
\arguments{
\item{openalex_id}{character compact OpenAlex ID or \code{NA}.}
}
\value{
full URL or \code{NA_character_}.
}
\description{
Build the full OpenAlex URL for a compact ID.
}
