% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations.R
\name{OpenCitation}
\alias{OpenCitation}
\title{Construct an \code{OpenCitation} data-carrier object.}
\usage{
OpenCitation(
  citing_doi = NA_character_,
  citing_pmid = NA_character_,
  cited_doi = NA_character_,
  creation_date = NA_character_,
  raw = list()
)
}
\arguments{
\item{citing_doi}{character (possibly \code{NA_character_})}

\item{citing_pmid}{character (possibly \code{NA_character_})}

\item{cited_doi}{character (possibly \code{NA_character_})}

\item{creation_date}{character ISO date (possibly \code{NA_character_})}

\item{raw}{list, the original record.}
}
\value{
An \code{OpenCitation} object.
}
\description{
A simple list-based S3 object holding the normalized fields of a single
OpenCitations record plus the raw source list.
}
