% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations.R
\name{parse_open_citation}
\alias{parse_open_citation}
\title{Parse a raw OpenCitation record list into an \code{OpenCitation} object.}
\usage{
parse_open_citation(item)
}
\arguments{
\item{item}{list with fields \code{citing}, \code{cited}, \code{creation}.}
}
\value{
an \code{OpenCitation} object.
}
\description{
Parse a raw OpenCitation record list into an \code{OpenCitation} object.
}
