% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{retractionpollution-package}
\alias{retractionpollution}
\alias{retractionpollution-package}
\title{retractionpollution}
\description{
Trace citation pollution from retracted and concerning scientific papers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahmad-sofimahmudi/RetractionPollution}
  \item Report bugs at \url{https://github.com/ahmad-sofimahmudi/RetractionPollution/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Sofi-Mahmudi \email{ahmad@sofimahmudi.com}

Authors:
\itemize{
  \item Ahmad Sofi-Mahmudi \email{ahmad@sofimahmudi.com}
}

}
\keyword{internal}
