% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{search_fallback_terms}
\alias{search_fallback_terms}
\title{Build fallback search terms from a seed.}
\usage{
search_fallback_terms(seed)
}
\arguments{
\item{seed}{named list (output of \code{row_to_seed}).}
}
\value{
named list with \code{title} and \code{author_last_name}.
}
\description{
Build fallback search terms from a seed.
}
