% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencitations.R
\name{stop_open_citations}
\alias{stop_open_citations}
\title{Raise an \code{OpenCitationsError}.}
\usage{
stop_open_citations(msg)
}
\arguments{
\item{msg}{character message.}
}
\value{
Never; signals an error.
}
\description{
Raise an \code{OpenCitationsError}.
}
