% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{text_or_none}
\alias{text_or_none}
\title{Normalize text, returning \code{NA_character_} for missing/empty input.}
\usage{
text_or_none(value)
}
\arguments{
\item{value}{any R object}
}
\value{
trimmed character string or \code{NA_character_}.
}
\description{
Normalize text, returning \code{NA_character_} for missing/empty input.
}
